ref(onboarding): Inline ScmView into scmConnect
No need for a separate component -- the connected view markup is small
and only used in one place.
